requirements.txt
streamlit>=1.28.0
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0